requirements.txt
Flask
Flask-Babel
python-dotenv
gunicorn  # Recommended for production
